refac/enh: move common classes from otbsuppl/OTBGDALRATImageIO to otbsuppl/core
Moving non IO classes out of OTBGDALRATImageIO library into OTBSupplCore.
This not only declutters OTBGDALRATImageIO, it also enables a better
link order of dependencies that was preventing netcdf_par to be included
properly by NMNetCDFIO and thus breaking parallel read/write of immages.
We also renamed the above mentioned IO libraries to indicate that
they are part of LUMASS to avoid confusing them with original OTB
or netcdf libraries.